+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 2.x branch
 ## 2.6 branch
 ### 2.6.4
-* now you can show banners and photos as list or as grid;
-* now you can download banners and photos.
+* you can show banners and photos as list or as grid;
+* you can filter banners and photos by created date;
+* you can download banners and photos.
 
 ### 2.6.3
 * fixed bug ordering posts and pages.